+++ HRD_App/Forms/WorkedTimes/WorkedTimesForm.cs
@@ -90,20 +90,24 @@ namespace HRD_App.Forms
         {
             if (workedTimes.Count < 1) return;
 
-            int row = dataGridView_workedTimes.CurrentRow.Index;
-            int id = (int)dataGridView_workedTimes[0, row].Value;
+            DialogResult result = MessageBox.Show("Вы действительно хотите удалить запись?", "", MessageBoxButtons.YesNo);
+            if (result == DialogResult.Yes)
+            {
+                int row = dataGridView_workedTimes.CurrentRow.Index;
+                int id = (int)dataGridView_workedTimes[0, row].Value;
 
-            try
-            {
-                WorkedTime workedTime = await RestApi.WorkedTimeService.Delete(id);
-                DeleteWorkedTime(workedTime);
-                Filter();
-            }
-            catch (Exception exception)
-            {
-                MessageBox.Show(exception.Message);
-                Console.WriteLine(exception.Message);
-                Console.WriteLine(exception.StackTrace);
+                try
+                {
+                    WorkedTime workedTime = await RestApi.WorkedTimeService.Delete(id);
+                    DeleteWorkedTime(workedTime);
+                    Filter();
+                }
+                catch (Exception exception)
+                {
+                    MessageBox.Show(exception.Message);
+                    Console.WriteLine(exception.Message);
+                    Console.WriteLine(exception.StackTrace);
+                }
             }
         }
